+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 ### Fixes
 - Fix cancellation ([#173](https://github.com/databricks/dbt-databricks/pull/173))
+- `http_headers` should be dict in the profile ([#174](https://github.com/databricks/dbt-databricks/pull/174))
 
 ## dbt-databricks 1.2.2 (September 8, 2022)